Bubble Tea
Saving progress on creating the window panes for the terminal app.
