+++ main.go
@@ -2,103 +2,80 @@ package main
 
 import (
 	"fmt"
+	"os"
+	"os/exec"
 
+	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/charmbracelet/lipgloss/tree"
 )
 
-type styles struct {
-	base,
-	block,
-	enumerator,
-	toggle,
-	db,
-	table,
-	column lipgloss.Style
+type model struct {
+	width  int
+	height int
+	ready  bool
 }
 
-func defaultStyles() styles {
-	var s styles
-	s.base = lipgloss.NewStyle().
-		Background(lipgloss.Color("57")).
-		Foreground(lipgloss.Color("225"))
-	s.block = s.base.
-		Padding(1, 3).
-		Margin(1, 3).
-		Width(40)
-	s.enumerator = s.base.
-		Foreground(lipgloss.Color("212")).
-		PaddingRight(1)
-	s.toggle = s.base.
-		Foreground(lipgloss.Color("207")).
-		PaddingRight(1)
-	s.db = s.base
-	s.table = s.base
-	s.column = s.base
-	return s
+func (m model) Init() tea.Cmd {
+	return nil
 }
 
-type db struct {
-	name   string
-	open   bool
-	styles styles
-}
+func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		switch msg.String() {
+		case "q":
+			return m, tea.Quit
+		}
+	case tea.WindowSizeMsg:
+		m.width = msg.Width
+		m.height = msg.Height
+		m.ready = true
 
-func (d db) String() string {
-	t := d.styles.toggle.Render
-	n := d.styles.db.Render
-	if d.open {
-		return t("▼") + n(d.name)
 	}
-	return t("▶") + n(d.name)
+
+	return m, nil
 }
 
-type table struct {
-	name   string
-	open   bool
-	styles styles
+func (m model) View() string {
+	style := lipgloss.NewStyle().Width(m.width).Height(m.height).Background(lipgloss.Color("63"))
+	return style.Render("This is a fullscreen application")
 }
 
-func (t table) String() string {
-	toggle := t.styles.toggle.Render
-	tableName := t.styles.table.Render
+// func (m model) View() string {
+// 	if !m.ready {
+// 		return "loading..."
+// 	}
+//
+// 	windowStyle := lipgloss.NewStyle().
+// 		Width(m.width/2).
+// 		Height(m.height/2).
+// 		Padding(1, 2).
+// 		Border(lipgloss.RoundedBorder()).
+// 		BorderForeground(lipgloss.Color("63")).
+// 		Align(lipgloss.Center)
+//
+// 	content := "Simple window"
+//
+// 	return windowStyle.Render(content)
+// }
 
-	if t.open {
-		return toggle("▼") + tableName(t.name)
-	}
-	return toggle("▶") + tableName(t.name)
-}
-
-type column struct {
-  name string
-  styles styles
-}
-
-func (c column) String() string {
-  return c.styles.column.Render(c.name)
-}
-
-func PostgresTreeView() {
-  s := defaultStyles()
-
-	t := tree.Root(db{"DB Name", true, s}).
-    Enumerator(tree.RoundedEnumerator).
-    EnumeratorStyle(s.enumerator).
-		Child("New Query").
-		Child("Saved Queries").
-		Child(
-			tree.New().
-				Root("Schemas").
-				Child("Information Schema").
-				Child("PG Catalog").
-				Child("PG Toast").
-				Child("Public"),
-		)
-
-	fmt.Println(s.block.Render(t.String()))
-}
-
-// Tree view using lipgloss
 func main() {
-	PostgresTreeView()
+	// save terminal state
+	saveState := exec.Command("tput", "smcup")
+	saveState.Stdout = os.Stdout
+	saveState.Run()
+
+	// Ensure the terminal state is restored when the program exits
+	defer func() {
+		restoreState := exec.Command("tput", "rmcup")
+		restoreState.Stdout = os.Stdout
+		restoreState.Run()
+	}()
+
+	p := tea.NewProgram(model{})
+
+	if err := p.Start(); err != nil {
+		fmt.Println("Error:", err)
+		os.Exit(1)
+	}
 }
